requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
click==8.1.7
fastapi==0.115.6
flake8==7.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
polars==1.17.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
pytest==8.3.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
python-dotenv==1.0.1

